% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyPeaksData.R
\name{DailyPeaksData}
\alias{DailyPeaksData}
\title{Extract Peaks from HYDAT database}
\usage{
DailyPeaksData(db, info, tol = 346)
}
\arguments{
\item{db}{Path to the HYDAT database.}

\item{info}{Site information. Must be a data frame with 3
columns: site, area, threshold.}

\item{tol}{Number of days considered as a complete year.}
}
\description{
Return a list of extracted independent thresholds.
}
\examples{

print('hello')


}
