% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData.R
\name{DemoData}
\alias{DemoData}
\title{Demonstration data}
\usage{
DemoData(type = "amax")
}
\arguments{
\item{type}{Type of hydrometric data to return.}
}
\description{
Return hydrometric data available in the package \code{CSHShydRology} but in the format
required by the modeling functions in \code{floodnetRfa}.
Different datasets can be requested:
the annual maxima from station 01AF007 (\code{'amax'}),
the daily streamflow (\code{'daily'}) or exceedances (\code{'peaks'})
from station 01AD002 and the annual maxima from 38 stations in the Atlantic
provinces (\code{'region'}).
}
\examples{

x <- DemoData()
head(x)

x <- DemoData(type = 'daily')
head(x)

x <- DemoData(type = 'peaks')
head(x)
PeaksMeta(x)

x <- DemoData(type = 'region')
head(x)

x <- DemoData(type = 'descriptor')
head(x)

}
\seealso{
\link[CSHShydRology]{flowStJohn}, \link[CSHShydRology]{flowAtlantic}.
}
