% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descriptors}
\alias{descriptors}
\title{Catchment descriptors}
\format{An object of class \code{data.frame} with 770 rows and 17 columns.}
\usage{
descriptors
}
\value{
The dataset includes the following variables:
\itemize{

  \item Station identification number: \code{station}.

  \item Station geographical coordinates: \code{lon, lat}.

  \item Watershed centers: \code{lon_ws, lat_ws}.

  \item Drainage area (m3/s): \code{area}.

  \item Perimeter (km): \code{peri}.

  \item Station elevation (m): \code{elev}.

  \item Watershed average elevation (m): \code{elev_ws}.

  \item Mean aspect (Degree): \code{aspect}.

  \item Mean slope (\%): (\code{slope}).

  \item Stream density (\eqn{km[-1]}): \code{stream}.

  \item Percentage of water bodies(\%): \code{wb}.

  \item Station and watershed mean annual precipitation (mm): \code{map, map_ws}.

  \item Station and watershed annual average air temperature: \code{temp,temp_ws}.

}
}
\description{
List of meteorological and physical characteristics of 770  gauged stations.
}
\references{
Durocher, M., Burn, D. H., & Mostofi Zadeh, S. (2018). A nationwide regional
flood frequency analysis at ungauged sites using ROI/GLS with copulas and
super regions. Journal of Hydrology, 567, 191-202.
https://doi.org/10.1016/j.jhydrol.2018.10.011
}
\keyword{datasets}
