% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetApp.R
\name{FloodnetApp}
\alias{FloodnetApp}
\title{Floodnet App}
\usage{
FloodnetApp(shortcut = NULL)
}
\arguments{
\item{shortcut}{File path of the launching script.}
}
\description{
A simple function used to start launch a Graphical User Interface (GUI)
inside a the default web browser.
Alternatively, if a path is passed in argument it create a script that will
launch the GUI from the system terminal.
}
\examples{

\dontrun{
  FloodnetApp()

  #On Windows
  floodnetApp("C:/Users/JohnDoe/Desktop/FloodnetApp.cmd")

  #On Unix
  floodnetApp("/home/JohnDoe/Desktop/FloodnetApp.sh")
}
}
