% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodnetGraphics.R, R/hist_floodnetMdl.R,
%   R/plot_floodnetQrt.R
\name{floodnetGraphics}
\alias{floodnetGraphics}
\alias{plot.floodnetMdl}
\alias{plot.floodnetMdls}
\alias{.PlotReturnLevel}
\alias{.PlotQQ}
\alias{.PlotCompareQ}
\alias{.PlotCompareCv}
\alias{.PlotLRatio}
\alias{.PlotLinTrend}
\alias{hist.floodnetMdl}
\alias{plot.floodnetRoi}
\title{Floodnet Plotting functions.}
\usage{
\method{plot}{floodnetMdl}(x, type = "r", ...)

\method{plot}{floodnetMdls}(x, type = "q", ...)

.PlotReturnLevel(
  x,
  point.args = NULL,
  line.args = NULL,
  ribbon.args = NULL,
  xlab = "Return periods",
  ylab = "Return levels"
)

.PlotQQ(
  x,
  point.args = NULL,
  line.args = NULL,
  ribbon.args = NULL,
  xlab = "Theoretical quantiles",
  ylab = "Sample quantiles"
)

.PlotCompareQ(x, xlab = "Return periods", ylab = "Return levels", fill = "")

.PlotCompareCv(
  x,
  line.args = NULL,
  point.args = NULL,
  xlab = "Return periods",
  ylab = "Coefficient of variation",
  colour = ""
)

.PlotLRatio(
  x,
  point.args = NULL,
  average.args = NULL,
  line.args = NULL,
  xlab = "L-Skewness",
  ylab = "L-Kurtosis",
  colour = "",
  shape = ""
)

.PlotLinTrend(
  x,
  point.args = NULL,
  line.args = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = TRUE
)

\method{hist}{floodnetMdl}(
  x,
  histogram.args = NULL,
  line.args = NULL,
  xlab = NULL,
  ylab = "Density",
  caption = TRUE,
  ...
)

\method{plot}{floodnetRoi}(x, type = "f", ...)
}
\arguments{
\item{x}{Input data.}

\item{type}{Plot type. \code{'r'} - Return level, \code{'q'} - compare quantile,
\code{'cv'} - compare cv, \code{'l'} - L-moment ratio.}

\item{...}{Other parameters.}

\item{point.args, line.args, histogram.args, ribbon.args, }{Parameters passed to the respective geometry.}

\item{xlab, ylab, fill, shape, colour}{Labels of the XY-axis or legend.}

\item{average.args}{Same as point.args for the average point.}

\item{caption}{Logical.
Should additional information be displayed in the caption.}
}
\description{
Generate diagnostic plots based on the output of a modeling function
\link{floodnetMdl}). Includes the following graphics: Return level plot,
comparison of the flood quantiles, comparison of the coefficient of
variation, L-moment ratio diagram and histogram.#'
}
\details{
All functions use \code{ggplot2} to create a basic plot that can be further
customized. The default arguments of the used geometry can be overridden by
passing a list of new arguments. For example,
\code{point.args = list(colour = 'red')}
change the colour of a set of points that was created by \code{geom_point}.
}
\examples{

library(ggplot2)

## Create a random data
set.seed(2)
xd <- SequenceData(365*30, freq = 'days', sdate = '2000-01-01')
an <- CSHShydRology::ExtractAmax(value~date,xd)

## Return level Plot
fit.an <- an$value \%>\%
  SequenceData() \%>\%
  FloodnetAmax()

plot(fit.an)

# QQ-plot
fit.pot <- FloodnetPot(xd, area = 200, u = 200)
plot(fit.pot, type = 'qq')

## Histogram
hist(fit.pot)

## Sen's slope
plot(fit.an, 't')

## Comparing flood quantiles and coefficient of variation
m2 <- CompareModels(fit.an, fit.pot)
plot(m2) + labs(fill = 'Method')
plot(m2, type = 'cv')

## L-moment ratio diagram
DemoData('region') \%>\%
  FloodnetPool(target = '01AF007', tol.H = Inf, verbose = FALSE) \%>\%
  plot(type = 'l') + labs(shape = 'Site', colour = 'Distr.')


}
\seealso{
\link[ggplot2]{ggplot}, \link[ggplot2]{geom_point},
  \link[ggplot2]{geom_path}, \link[ggplot2]{geom_ribbon}
}
