% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModel.R, R/as_data_frame_floodnetMdl.R,
%   R/as_data_frame_floodnetQrt.R, R/floodnetMdl.R
\name{CompareModels}
\alias{CompareModels}
\alias{as.data.frame.floodnetMdl}
\alias{as.data.frame.floodnetMdls}
\alias{as.data.frame.floodnetRoi}
\alias{floodnetMdl}
\alias{summary.floodnetMdl}
\alias{summary.floodnetMdls}
\alias{print.floodnetMdl}
\alias{print.floodnetRoi}
\alias{summary.floodnetRoi}
\title{Floodnet model output}
\usage{
CompareModels(...)

\method{as.data.frame}{floodnetMdl}(x, row.names = NULL, optional = FALSE, type = "q", ...)

\method{as.data.frame}{floodnetMdls}(x, row.names = NULL, optional = FALSE, type = "q", ...)

\method{as.data.frame}{floodnetRoi}(x, row.names = NULL, optional = FALSE, type = "q", ...)

\method{summary}{floodnetMdl}(object, ...)

\method{summary}{floodnetMdls}(object, ...)

\method{print}{floodnetMdl}(x, ...)

\method{print}{floodnetRoi}(x, ...)

\method{summary}{floodnetRoi}(object, print.it = TRUE, ...)
}
\arguments{
\item{...}{Other parameters.}

\item{x, object}{Output of a modeling function. Such as \link{FloodnetAmax}.}

\item{row.names}{Optional vector of names for the rows.}

\item{optional}{Logical. Should \link[base]{make.names} be applied to row names.}

\item{type}{Type of output to be returned:
parameter \code{'p'} or quantile \code{'q'}.}

\item{print.it}{Logical. Should the results be displayed.}
}
\description{
All Floodnet functions create a standardized model output that can be used
for printing results or generating graphics.
}
\seealso{
\link{FloodnetAmax},\link{FloodnetPot}, \link{FloodnetPool},
  \link{floodnetGraphics}, \link{hist.floodnetMdl}.
}
