% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPool.R, R/FloodnetPoolMle.R
\name{FloodnetPool}
\alias{FloodnetPool}
\alias{FloodnetPoolMle}
\title{Pool frequency analysis using annual maximum floods}
\usage{
FloodnetPool(x, target, period = c(2, 5, 10, 20, 50, 100),
  distr = NULL, tol.H = 2, nsim = 2000, alpha = 0.05,
  corr = NULL, out.model = FALSE, verbose = TRUE)

FloodnetPoolMle(x, target, period = c(2, 5, 10, 20, 50, 100), distr,
  type = "mean", nsim = 2000, alpha = 0.05, corr = NULL,
  out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{Hydrometric data of the form: site, year, value.}

\item{target}{Station ID of the target.}

\item{period}{Return periods for which the flood quantiles are estimated.}

\item{distr}{Regional distribution.}

\item{tol.H}{Heterogeneity measure. Stopping criterion representing the
minimal heterogeneity level accepted for a pooling group.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{alpha}{Probability outside the confidence interval.}

\item{corr}{Coefficient of correlation used for simulation.
The same value is assumed between each pair of sites.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitRegLmom}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be returned.}

\item{type}{Type of regional models to be fitted.
See \link[CSHShydRology]{FitPoolMle}.}
}
\description{
Return the results of a regional flood frequency analysis based on annual
maxima and pooling groups.
}
\details{
Estimation is carried out by the L-moment algoritms.
If not provided the distance between sites is evaluated using the distance
between the regularity and timing of the annual flood peaks.
Confidence intervals and standard deviation are evaluated by a
parametric bootstrap.
An empirical matrix of intersite correlation is estimated and used to simulate
from a multivariate normal distribution.
}
\examples{

\dontrun{
## Assuming that a downloaded version of HYDAT database is available
db <- "extdata/Hydat.sqlite3"
target <- '01AF009'
supreg <- with(gaugedSites, station[supreg_km12 == 11])

an <- AmaxData(sites = supreg, target = target, db = db)

FloodnetPoolAmax(an, target, period = c(20,50))
}

}
