% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPoolAmax.R
\name{FloodnetPoolAmax}
\alias{FloodnetPoolAmax}
\title{Pool frequency analysis using annual maximum floods}
\usage{
FloodnetPoolAmax(period = c(2, 5, 10, 20, 50, 100), target,
  sites = NULL, db = NULL, x = NULL, distr = NULL, size = 25,
  tol.H = 2, nsim = 2000, alpha = 0.05, out.model = FALSE,
  verbose = TRUE)
}
\arguments{
\item{period}{Return periods for which the flood quantiles are estimated.}

\item{target}{Station ID for HYDAT.}

\item{sites}{List of station ID for HYDAT that may be used
for the regional analysis.}

\item{db}{Path of the HYDAT database.}

\item{x}{Path of a CSV file or data.frame containing the flow data.}

\item{distr}{Regional distribution.}

\item{size}{Initial size of the pooling groups.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{alpha}{Probability outside the confidence interval.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitRegLmom}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be returned.}
}
\description{
Return the results of a regional flood frequency analysis based on annual
maxima and pooling groups.
Estimation is carried out by the L-moment algoritms.
}
\examples{

\dontrun{
## Assuming that the HYDAT database is available
db <- "/pathToDB/HYDAT.sqlite"
FloodnetPoolAmax(period = c(20,50), target = '01AD002', db = db)
}

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
