% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPoolPot.R
\name{FloodnetPoolPot}
\alias{FloodnetPoolPot}
\title{Pool frequency analysis using peaks over threshold}
\usage{
FloodnetPoolPot(period = c(2, 5, 10, 20, 50, 100), target, sites,
  db = NULL, x = NULL, nsim = 2000, alpha = 0.05, size = 25,
  tol.H = 2, tol.year = 346, out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{period}{Return periods for which the flood quantiles are estimated.}

\item{target}{Station ID for HYDAT.}

\item{sites}{List of station ID for HYDAT that may be used
for the regional analysis. Must be a data.frame with three columns.
Respectively:  site, threshold and drainage area.}

\item{db}{Path of the HYDAT database.}

\item{x}{Path of a CSV file or data.frame containing the flow data.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{alpha}{Probability outside the confidence interval.}

\item{size}{Initial size of the pooling groups.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitRegLmom}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be returned.}
}
\description{
Return the results of a regional flood frequency analysis based on
peaks over threshold and pooling groups.
Estimation is carried out by the L-moment algoritms.
}
\examples{

\dontrun{
## Assuming that the HYDAT database is available
db <- "/pathToDB/HYDAT.sqlite"
FloodnetPoolPot(period = c(20,50), target = '01AD002', db = db)
}

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
