% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPot.R
\name{FloodnetPot}
\alias{FloodnetPot}
\title{Estimate flood quantiles using using peaks over threshold}
\usage{
FloodnetPot(period = c(2, 5, 10, 20, 50, 100), x = NULL, site = NULL,
  db = NULL, u = NULL, area = NULL, nsim = 2000, alpha = 0.05,
  tol.year = 346, out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{period}{Return period for which the flood quantiles are estimated.}

\item{x}{Path of a CSV file or data.frame.}

\item{site}{Station ID for HYDAT or a station name for identification.}

\item{db}{File name of the HYDAT database.}

\item{u}{Threshold value provided by the user.}

\item{area}{Drainage area.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{alpha}{Probability outside the confidence interval.}

\item{tol.year}{Number of days necessary to consider a year complete.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitPot}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be output.}
}
\description{
Return the flood quantiles estimated by (at-site) Peaks Over Threshold (POT).
Estimation is carried out by maximum likelihood using the Generalized
Pareto distribution. If not provided, a best candiate threshold is searched.
Apart some restrictions, it corresponds to the lowest
threshold that passes the goodness of fit test of Anderson-Darling with
a p-value greater that 0.25.
In addition, the selected threshold is forced to have in
average less than 2.5 peaks per year and at least 30 peaks. If
no threshold is associated to p-value greater that 0.25, the threshold
with the maximimum p-value is used instead.
The declustering technique for extracting the independend peaks
follows the recommendations of the Water Ressources council of United States.
In particular, peaks must be separated by at least 4 + log(A) days, where
A is the drainage area of the basin in square kilometers.
Inference on the estimated flood quantiles is performed by parametric bootstrap.
}
\details{
If the HYDAT database is used as input, the the drainage area
is extracted from the database. If no drainage area is provided,
its value is approximated by a log-log relationship with mean river
discharge. This relationship was estimated using the 1114 sites found in the
dataset `gaugedSites`.
It has the form $log(A) <- 4.0934 + 0.9944 * log(M)$, where M is the mean
streamflow.

For the CSV file, there must be at least two columns: Date and Value. The
names of the variables are not case sensitive. On the website of the
Water Survey of Canada, it is suggested to download the format
`Daily Data: Date-Data Format`.
}
\examples{

\dontrun{
## Assuming that the HYDAT database was already downloaded
db <- "/pathToDB/HYDAT.sqlite"
FloodnetPot(period = c(20,50), site = '01AD003', db = db)
}

}
