% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetRoi.R
\name{FloodnetRoi}
\alias{FloodnetRoi}
\title{Prediction of flood quantiles at ungauged sites using Region of Influence}
\usage{
FloodnetRoi(target, target.coord = NULL, sites, sites.coord = NULL,
  db = NULL, x = NULL, size = 20, period = 100, distr = "gev",
  nsim = 0, out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{target}{Descriptors of the target sites. A data.frame where the first
column is the name of the basin.}

\item{target.coord}{Coordinates of the target sites.}

\item{sites}{Descriptors of gauged sites. A data.frame where the first
column is the name of the basin.}

\item{sites.coord}{Coordinates of the gauged sites.}

\item{db}{The HYDAT database.}

\item{x}{Dataset containing the hydrometric data.}

\item{size}{Size of the region of influence.}

\item{period}{Return period to predict. Must be a single value.}

\item{distr}{Distribution of the gauged sites. Can a common distribution or
a vector of the individual distributions.}

\item{nsim}{Number of bootstrap samples.}

\item{out.model}{Logical. Should the model be returned.}

\item{verbose}{Should progress and message be display during the procedure.}
}
\description{
Return the flood quantiles, root mean square error, relative bias, lower and upper
bound of the 95% confidence interval estimated provided with a combination of
local regression and kriging technique.
}
\details{
The function uses a quantile regression technique (QRT) to predict flood
quantiles of a given return period at ungauged sites.
The function first perform at-site flood frequency analysis of the annual
maximum discharges extracted from the HYDAT database (or provided data).
At-site estimates of the flood quantile are evaluated by the L-moments method.
The resulting flood quantiles are fed to the QRT model to predict the
flood quantiles at ungauged sites according to its basin characteristics.
The QRT model uses a local regression method to evaluate flood quantiles at
specific target locations.
If coordinates are provided, simple kriging is additionally used to improve
the prediction by extracting further information
from the spatially correlated residuals.

If the size of the region of influence (ROI) is not provided,
a value is automatically selected.
If the dataset include fewer than 30 sites, all sites will beincluded by default,
otherwise ROI sizes of 20 and more are tried by steps of 5. The one
associated with the lowest mean absolute prediction error
(based on leave-one-out cross-validation) is finally selected

Please note that the distance between sites is the Euclidean distance and
that proper projection of the coordinates is required for input.

A mix of parametric bootstraps for the at-site distributions and residual
bootstraps for the QRT model is used to evaluate model uncertainty.
}
\examples{

\dontrun{
db <- "/pathToDB/HYDAT.sqlite"

}

}
\references{
Durocher, M., Burn, D. H., & Mostofi Zadeh, S. (2018). A nationwide regional
  flood frequency analysis at ungauged sites using ROI/GLS with copulas and
  super regions. Journal of Hydrology, 567, 191–202.
  https://doi.org/10.1016/j.jhydrol.2018.10.011

Durocher, M., Burn, D. H., Zadeh, S. M., & Ashkar, F. (2019). Estimating
  flood quantiles at ungauged sites using nonparametric regression methods
  with spatial components. Hydrological Sciences Journal, 64(9), 1056–1070.
  https://doi.org/10.1080/02626667.2019.1620952
}
