% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetToolbox.R
\name{FloodnetToolbox}
\alias{FloodnetToolbox}
\title{Run Shiny Toolbox GUI}
\usage{
FloodnetToolbox()
}
\value{
None
}
\description{
A simple function used to run the Shiny GUI from the floodnetRfa package isntallation
}
\examples{
FloodnetToolbox()
}
