% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSuperRegion.R
\name{GetSuperRegion}
\alias{GetSuperRegion}
\title{Super regions}
\usage{
GetSuperRegion(
  site,
  type = "amax",
  cluster = "km12",
  trend.tol = 0.05,
  thresh = "auto"
)
}
\arguments{
\item{site}{Station identification number.}

\item{type}{Type of frequency analysis: Annual maximum (\code{'amax'}) or
peaks over threshold (\code{'pot'}).}

\item{cluster}{Clustering thechniques. It corresponds to a column of
\code{gaugedSites}. Must be one of \code{'km6'}, \code{'km12'},
\code{'hc6'} or \code{'hc12'}.}

\item{trend.tol}{Significance level of the trend tests. This used to remove
stations that may not have constant flood risk over time.}

\item{thresh}{Threshold to return. It corresponds to a column of
\code{gaugedSites}. For example, \code{'auto'} or \code{'ppy200'}.}
}
\description{
Return the super region associated to a target site as suggested by the
information provided in the dataset \code{gaugedSites}.
}
\examples{

GetSuperRegion('01AF009')
GetSuperRegion('01AD002', type = 'pot', thresh = 'ppy200')

}
\seealso{
\link{gaugedSites}
}
