% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersiteCorData.R
\name{IntersiteCorData}
\alias{IntersiteCorData}
\title{Intersite correlation matrix}
\usage{
IntersiteCorData(db, x)
}
\arguments{
\item{db}{HYDAT database}

\item{x}{List of stations}
}
\description{
Return the intersite correlation matrix list of station based on annual
maximum discharge extracted from HYDAT.
}
\details{
The intersite correlations represent the correlation coefficients of an underlying
multivariate Gaussian distribution (Normal Copula).
First the spearman correlation is evaluated for each pair of sites with a
distance less than 500km.
Next the weighted least-squared method is used to evaluate the
parameter of an exponential model.
Finally, the fitted coefficient of each sites is returned.
}
\examples{

\dontrun{
##Intersite correlation for a given super region
IntersiteData(DB_HYDAT, GetSuperRegion('01AF009'))
}
}
\references{
Durocher, M., Burn, D. H., & Mostofi Zadeh, S. (2018). A nationwide regional
  flood frequency analysis at ungauged sites using ROI/GLS with copulas and
  super regions. Journal of Hydrology, 567, 191–202.
  https://doi.org/10.1016/j.jhydrol.2018.10.011
}
