% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_ca}
\alias{map_ca}
\title{Map of Canada}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 13 rows and 83 columns.
}
\usage{
map_ca
}
\description{
To produce a quick and simple map of Canada.
Original data were from https://www.naturalearthdata.com/.
}
\keyword{datasets}
