% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCA.R
\name{MapCA}
\alias{MapCA}
\title{Simple map of Canada}
\usage{
MapCA(..., polygon.args = NULL, xlab = "Longitude", ylab = "Latitude")
}
\arguments{
\item{...}{Other parameters passed to \link[ggplot2]{ggplot}.}

\item{polygon.args}{Parameters of the displayed polygons.
See \link[ggplot2]{geom_polygon}.}

\item{xlab, ylab}{Labels of the axes.}
}
\description{
Return a simple map of Canada using the grammar of graphics \code{ggplot2}.
}
\examples{

library(ggplot2)

MapCA(polygon.args = list(fill = 'white', colour = 'black'))

xd <- DemoData('descriptor')

MapCA() +
  geom_point(data = xd, aes(x = lon, y = lat), colour = 'orange') +
  coord_cartesian(xlim = c(-70,-58), ylim = c(43,50))

}
\seealso{
\link{floodnetGraphics}, \link[ggplot2]{geom_polygon}.
}
