% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonDistanceData.R
\name{SeasonDistance}
\alias{SeasonDistance}
\alias{SeasonDistanceData}
\title{Similarity measure based on seasonality of annual maxima}
\usage{
SeasonDistance(x, form = date ~ site, target = NULL, ...)

SeasonDistanceData(db, sites, ...)
}
\arguments{
\item{x}{Hydrometric data.}

\item{form}{Formula of the form date~site, that point to the columns with
date and site ID.}

\item{target}{Sites for which distance is returned.}

\item{...}{Other parameter passed to \link[CSHShydRology]{DistSeason}.}

\item{db}{Path of the HYDAT data base,}

\item{sites}{List of sites in HYDAT.}
}
\description{
Return the distance matrix between stations based on the
seasonality of the annual maximum events.
Optionally, a target can be passed to extract only the
pooling group from a list of stations.
}
\examples{

## Extract similarity measure from multiple site
## with respect to target station 01AF009
h <- SeasonDistance(DemoData('region'))
print(h[1:3,1:3])
print(h[2,])

\dontrun{

## Path to HYDAT
# DB_HYDAT = ".../hydat.sqlite3"

## Compute the distance using HYDAT directly
SeasonDistanceData(DB_HYDAT, names(h), target = '01AF009')

}

}
\seealso{
\link{AmaxData}.
}
