% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonDistanceData.R
\name{SeasonDistanceData}
\alias{SeasonDistanceData}
\title{Seasonal distance between stations}
\usage{
SeasonDistanceData(sites, db, x = NULL, target = NULL, ...)
}
\arguments{
\item{sites}{Stations}

\item{db}{HYDAT database}

\item{x}{Alternative input. Must be a data.frame of annual maximums
with 3 columns: site, time, value. The time variable must be of the class 'Date'.}

\item{...}{Other parameter passed to \link[CSHShydRology]{DistSeason}.}
}
\description{
Return the distance matrix between stations based on the
annual flood peaks regularity and timing.
}
\examples{
\dontrun{
db <- 'extdata/Hydat.sqlite3'

mysites <- c("01AD002", "01AD003", "01AE001", "01AF007", "01AF009")

SeasonDistanceData(mysites, db)
}

}
