% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData.R
\name{SequenceData}
\alias{SequenceData}
\title{Prepare hydrometric data from outside sources}
\usage{
SequenceData(
  value = 50,
  date = NULL,
  site = "site1",
  freq = "years",
  sdate = Sys.Date()
)
}
\arguments{
\item{value}{Hydrometric data.}

\item{date}{Date.}

\item{site}{Site identification number.}

\item{freq}{Frequency of the input data.}

\item{sdate}{Starting date.}
}
\description{
Prepare a dataset having three columns (site, date, value) from vectors of
sequential data. If only an integer is passed as input, a sample of the same
size is simulated from a Gumbel distribution.
The data has a mean equal to 100 and a standard deviation
equal to 30.
}
\examples{

## Example of a yearly sequence, starting today.
mydata <- rnorm(3)
SequenceData(mydata, site = 'mysite')


## Example of daily sequence starting in 2000-01-01.
SequenceData(mydata, freq = 'days', sdate = '2000-01-01')


## Example of an unregular sequence.
mydate <- as.Date(c('2000-01-01','2005-05-05','2010-10-10'))
SequenceData(mydata, mydate)

## Create a testing set
SequenceData(3, site = c('s1','s2'))

}
\seealso{
\link{AmaxData}, \link{DemoData}.
}
