% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationData.R
\name{StationData}
\alias{StationData}
\title{Station information}
\usage{
StationData(db, sites)
}
\arguments{
\item{db}{HYDAT database.}

\item{sites}{List of stations.}
}
\description{
Return the coordinates and drainage area found in the HYDAT database
 for a group of stations.
}
\examples{

\dontrun{

StationData(DB_HYDAT, c('01AD002','01AF009'))

}

}
