% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmaxData.R, R/DailyData.R
\name{AmaxData}
\alias{AmaxData}
\alias{DailyData}
\title{Extract Annual, daily data from HYDAT}
\usage{
AmaxData(sites, db, year = TRUE, target = NULL, size = 25, distance = NULL)

DailyData(
  sites,
  db,
  pad = FALSE,
  tol = 346,
  target = NULL,
  size = 25,
  distance = NULL
)
}
\arguments{
\item{sites}{List of stations.}

\item{db}{HYDAT database.}

\item{year}{Logical. Should the full date be return or only the year.}

\item{target}{A target site of the pooling group.}

\item{size}{Size of the pooling group.}

\item{distance}{A vector of distance between all sites.}

\item{pad, tol}{Logical and number of days. Should the daily data be padded.
See \link[CSHShydRology]{PadPot}.}
}
\description{
Return a dataset of hydrometric data from HYDAT.
Optionally, a target site can be passed and so only its pooling groups is returned.
}
\examples{

\dontrun{
## This create a variable DB_HYDAT that point to database
db <- DB_HYDAT

## Reading AMAX data for one station
x <- AmaxData(c('01AD002'), db)
head(x, 3)

## Reading multiple stations
x <- AmaxData(c('01AD002','01AF009'), db, year = FALSE)
x[seq(85,95),]

## Reading Daily data
x <- DailyData(c('01AD002','01AF009'), db)
head(x)

## Filter the stations to keep only a pooling group of size 5
sid <- gaugedSites$supreg_km12 == 11
sites <- gaugedSites$station[sid]

coord <- gaugedSites[sid, c('lon','lat')]
rownames(coord) <- sites

h <- as.matrix(dist(coord))

x <- AmaxData(sites, db, target = '01AF009', size = 5,  distance = h)
table(x$station)
round(sort(h['01AF009',])[1:10],2)
}

}
\seealso{
\link{DailyPeaksData}.
}
