% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmaxData.R, R/DailyData.R
\name{AmaxData}
\alias{AmaxData}
\alias{DailyData}
\title{Extract annual or daily hydrometric data from HYDAT}
\usage{
AmaxData(db, sites, target = NULL, distance = NULL, size = 25)

DailyData(
  db,
  sites,
  pad = FALSE,
  tol = 346,
  target = NULL,
  distance = NULL,
  size = 25
)
}
\arguments{
\item{db}{Path of the HYDAT database.}

\item{sites}{List of stations from which to extract the data.}

\item{target}{Target site.}

\item{distance}{A vector of distance between all sites. The target has distance
equal to zero.}

\item{size}{Size of the pooling group.}

\item{pad, tol}{Logical and number of days. Should the daily data be padded.
See \link[CSHShydRology]{PadPot}.}
}
\description{
Return a dataset of hydrometric data from a local HYDAT database.
Optionally, a target or vector of distances can be passed to extract only the
pooling group from a list of stations.
If a distance is passed, the target has a distance of zero.
If a target is passed, the similarity measure based on the seasonality of the
annual maxima are used as the distance.
}
\examples{

\dontrun{
## The variable DB_HYDAT containing the path to the HYDAT database.
#DB_HYDAT <- '.../Hydat.sqlite3'
#gaugedSites <- read.csv('.../gauged_sites.csv')

## Reading AMAX data for one station.
x <- AmaxData(DB_HYDAT, c('01AD002'))
head(x, 3)

## Reading multiple stations.
x <- AmaxData(DB_HYDAT, c('01AD002', '01AF009'))
x[seq(85,95),]

## Reading Daily data.
x <- DailyData(DB_HYDAT, c('01AD002','01AF009'))
head(x, 3)

## A pooling group of size 5 based on seasonality distance.
x <- AmaxData(DB_HYDAT, gaugedSites$station, target = '01AF009', size = 5)

## Extracted site.
sort(unique(x$site))

## Pooling group with different a distance.
meta <- log(gaugedSites[, c('area','map')])
h <- as.matrix(dist(scale(meta)))
x <- AmaxData(DB_HYDAT, gaugedSites$station, distance = h[2,], size = 5)

## Extracted site.
sort(unique(x$site))

}

}
\seealso{
\link{DailyPeaksData}, \link{SeasonDistance}.
}
