% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyPeaksData.R, R/ExtractPeaksData.R,
%   R/PeaksData.R
\name{DailyPeaksData}
\alias{DailyPeaksData}
\alias{ExtractPeaksData}
\alias{PeaksData}
\alias{PeaksMeta<-}
\alias{PeaksMeta}
\title{Extract independent Peaks}
\usage{
DailyPeaksData(
  db,
  info,
  target = NULL,
  size = 25,
  distance = NULL,
  pad = FALSE,
  tol = 346
)

ExtractPeaksData(x, info, pad = FALSE, tol = 346, sorted = FALSE)

PeaksData(x, meta)

PeaksMeta(x) <- value

PeaksMeta(x)
}
\arguments{
\item{db}{Path to the HYDAT database.}

\item{info}{Site information. Must have 3 columns:
site, threshold and drainage area.}

\item{target}{Target station of the pooling group.}

\item{size}{Size of the pooling group.}

\item{distance}{Distance between stations. If not provided, the distance
between the regularity and timing of the annual flood peaks is used.}

\item{pad}{Logical. Should the time series be padded.
See \link[CSHShydRology]{PadPot}.}

\item{tol}{Number of days to consider a year as complete.}

\item{x}{Hydrometric Data. Must have 3 columns: site, date and value.}

\item{sorted}{Logical. Is the data sorted.}

\item{meta, value}{Meta information about thresholds.
Must be a data frame with 3
columns: site, threshold and exceedance rates (peaks per year).}

\item{...}{Other parameters}
}
\description{
The function `ExtractPeaksData` returned independent peaks for one or more
sites and the function `DailyPeaksData` applies it directly on daily data
extracted from HYDAT.
The function `PeaksMeta` serves to link meta-information about the threshold
and exceedance rate to the extracted peaks.
}
\examples{

\dontrun{
 ## The variable DB_HYDAT contains the path to the HYDAT database.
 DB_HYDAT <- '.../Hydat.sqlite3'

 ## Meta-information about the stations
 gaugedSites <- read.csv('.../gauged_sites.csv')

 ## Data.frame containing thresholds and drainage area
 info <- gaugedSites[1:2, c('station','ppy200','area')]

 ## Reading for one station
 x <- DailyPeaksData(DB_HYDAT, info)
 head(x)

 ## Manually extracting the data
 DB_HYDAT \%>\%
  DailyData(info$station) \%>\%
   ExtractPeaksData(info) \%>\%
   head()

 ## Create a dataset of exceedances manually
 xd <- SequenceData(3, site = unique(info$station))
 PeaksMeta(xd) <- info

}

}
\seealso{
\link{DailyData}, link{FloodnetPool}.
}
