% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindPeaksData.R, R/DailyPeaksData.R,
%   R/ExtractPeaksData.R, R/PeaksData.R
\name{BindPeaksdata}
\alias{BindPeaksdata}
\alias{DailyPeaksData}
\alias{ExtractPeaksData}
\alias{PeaksData}
\title{Extract Peaks from HYDAT database}
\usage{
BindPeaksdata(...)

DailyPeaksData(info, db, pad = FALSE, tol = 346, target = NULL,
  size = 25, distance = NULL)

ExtractPeaksData(x, info, pad = FALSE, tol = 346, sorted = FALSE)

PeaksData(x, info)
}
\arguments{
\item{...}{Other parameters}

\item{info}{Site information. Must be a data frame with 3
columns: site, threshold and area, except for function \code{PeaksData} where
area must be the total number of years.}

\item{db}{Path to the HYDAT database.}

\item{pad}{Logical. Should the time series be padded.
See \link[CSHShydRology]{PadPot}.}

\item{tol}{Number of days considered as a complete year.}

\item{target}{Target station of the pooling group.}

\item{size}{Size of the pooling group.}

\item{distance}{Distance between stations. If not provided, the distance
between the regularity and timing of the annual flood peaks is used.}

\item{x}{Hydrometric Data. Must have 3 columns:
station, threshold and drainage area.}

\item{sorted}{Logical. Is the data sorted.}
}
\description{
The function `DailyPeaksData` return an object containing exccedances above
given thresholds extracted from HYDAT.
The function `ExtractPeaksData` extract the same object from a data.frame.
The function `PeaksData` serves to create the same output object from known
exceedances and `BindPeaksData` can merged together 2 of these objectes.
}
\details{
For \code{DailyPeaksData}, if \code{info} has two columns, they are assumed
to be the station and threshold.
The drainage area is then extracted from the HYDAT database.

The utility function \code{PeaksData} can be used to construct the same output
from already extracted peaks.
}
\examples{

\dontrun{
 ## Path the HYDAT database
 db <- DB_HYDAT

 ## Data.frame containing the threshold and drainage area
 info <- gaugedSites[1:2, c('station','ppy200','area')]

 ## Reading AMAX data for one station
 x <- DailyPeaksData(info, db, pad = TRUE)
 head(x$peaks, 3)

 ## Manually extracting the data
 xd <- DailyData(info$station,db)
 x2 <- ExtractPeaksData(xd, info, pad = TRUE)
 head(x2$peaks, 3)

 ## Create and merged Peaks
 p1 <- PeaksData(x2$peaks, info)
 p2 <- BindPeaksdata(p1, p1)

}

}
\seealso{
\link{DailyData}.
}
