% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindPeaksData.R, R/DailyPeaksData.R,
%   R/ExtractPeaksData.R, R/PeaksData.R
\name{BindPeaksdata}
\alias{BindPeaksdata}
\alias{DailyPeaksData}
\alias{ExtractPeaksData}
\alias{PeaksData}
\title{Extract Peaks from HYDAT database}
\usage{
BindPeaksdata(...)

DailyPeaksData(info, db, pad = FALSE, tol = 346, target = NULL,
  size = 25, distance = NULL)

ExtractPeaksData(x, info, pad = FALSE, tol = 346, sorted = FALSE)

PeaksData(x, info)
}
\arguments{
\item{...}{Other parameters}

\item{info}{Site information. Must be a data frame with 3
columns: site, threshold and area, except for function \code{PeaksData} where
area must be the total number of years.}

\item{db}{Path to the HYDAT database.}

\item{pad}{Logical. Should the time series be padded.
See \link[CSHShydRology]{PadPot}.}

\item{tol}{Number of days considered as a complete year.}

\item{target}{Target station of the pooling group.}

\item{size}{Size of the pooling group.}

\item{distance}{Distance between stations. If not provided, the distance
between the regularity and timing of the annual flood peaks is used.}

\item{x}{Hydrometric Data. Must have 3 columns:
station, threshold and drainage area.}

\item{sorted}{Logical. Is the data sorted.}
}
\description{
Return a list of peaks over thresholds data for one or more stations.
}
\details{
For \code{DailyPeaksData}, if \code{info} has two columns, they are assumed
to be the station and threshold.
The drainage area is then extracted from the HYDAT database.

The utility function \code{PeaksData} can be used to construct the same output
from already extracted peaks.
}
\examples{

\dontrun{
## A copy of HYDAT must be installed. Here an example
db <- 'extdata/Hydat.sqlite3'

sites <- with(gaugedSites, station[supreg_km12 == 11])

info <-  with(gaugedSites, gaugedSites[station \%in\% sites,
                                       c('station','auto','area')])

out <- DailyPeaksData(info, db, target = '01AF009', size = 5)
}

}
