% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descriptors}
\alias{descriptors}
\title{Station characteristics}
\format{An object of class \code{data.frame} with 770 rows and 17 columns.}
\usage{
descriptors
}
\description{
List of meteorological and physical characteristics for 770  gauged stations.
}
\details{
The dataset includes the following variables:
Station number (\code{station}),
Station geographical coordinates (\code{lon,lat}),
Coordinates of the center of the watershed (\code{lon_ws,lat_ws}),
Drainage area (\code{area}) and perimeter(\code{peri}) of the watershed,
Station elevation (\code{elev}) and average elevation of the watershed
(\code{elev_ws}),
Mean watershed aspect in Degree (\code{aspect}),
Mean slope in percent (\code{slope}),
Stream density (\code{stream}),
Percentage of water bodies (\code{wb}),
Mean annual precipitation for the station and watershed (\code{map,map_ws}) and
Average annual air temperature for the station and watershed (\code{temp,temp_ws}).

Durocher, M., Burn, D. H., & Mostofi Zadeh, S. (2018). A nationwide regional
flood frequency analysis at ungauged sites using ROI/GLS with copulas and
super regions. Journal of Hydrology, 567, 191-202.
https://doi.org/10.1016/j.jhydrol.2018.10.011
}
\keyword{datasets}
