% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetAmax.R
\name{FloodnetAmax}
\alias{FloodnetAmax}
\title{Estimate flood quantiles using using annual maxima}
\usage{
FloodnetAmax(period = c(2, 5, 10, 20, 50, 100), x = NULL,
  site = NULL, db = NULL, distr = NULL, instant = FALSE,
  nsim = 2000, alpha = 0.05, out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{period}{Return period for which the flood quantiles are estimated.}

\item{x}{Path of a CSV file or data.frame.}

\item{site}{Station ID for HYDAT or a station name for identification.}

\item{db}{Path of the HYDAT database.}

\item{instant}{Logical, should the instantaneous peaks be used.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{alpha}{Probability outside the confidence interval.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitAmax}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be returned.}

\item{maximum}{Logical. Should the maxima (or minima) be modeled.}
}
\description{
Return the results of an at-site frequency analysis of floods based on annual
maxima. Estimation is carried out by L-moments and a best distribution is
selected using Akaike Information Criterion (AIC) among the
Generalized Extreme Value (GEV), Generalized logistic (GLO), the
Generalized Normal (GNO) and the Pearson type 3 distributions. A preference is
given to the GEV when the AIC of the other distributions is not
greater than at least two points.
Inference on the estimated flood quantiles is performed by parametric bootstrap.
}
\details{
If `x` is a path, it is assumed to point to a CSV file, otherwise
it must be a data.frame.
In both cases, there must be at least one column
called `Max` (or `Min` if low flow are modeled).
The name of each column is not case sensitive.
On the website of the Water Survey of Canada, it is suggested to
download the the Annual Extremes Data.
}
\examples{

\dontrun{
## Assuming that the HYDAT database was already downloaded
db <- "/pathToDB/HYDAT.sqlite"
FloodnetAmax(period = c(20,50), site = '01AD002', db = db)
}

}
\author{
Martin Durocher <mduroche@uwaterloo.ca>
}
