% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetAmax.R
\name{FloodnetAmax}
\alias{FloodnetAmax}
\title{Estimate flood quantiles using using annual maxima}
\usage{
FloodnetAmax(x, ...)
}
\arguments{
\item{x}{Hydrometric data. Vector of annual maxima.}

\item{period}{Return period for which the flood quantiles are estimated.}

\item{distr}{Distribution. One of \code{'gev'},\code{'gno'},\code{'glo'}
or \code{'pe3'}}

\item{instant}{Logical, should the instantaneous peaks be used.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{level}{Confidence level.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitAmax}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be returned.}
}
\description{
Return the flood quantile, standard deviation,
of an at-site frequency analysis of floods based on annual
maxima.
}
\details{
Estimation is carried out by L-moments and a best distribution is
selected using Akaike Information Criterion (AIC) among the
Generalized Extreme Value (GEV), Generalized logistic (GLO), the
Generalized Normal (GNO) and the Pearson type 3 distributions.
A preference is given to the GEV when the AIC of the other distributions is not
greater than at least two.
Inference on the estimated flood quantiles is performed by parametric bootstrap.

If \code{verbose == TRUE}, a Mann-Kendall and a Pettitt's test are used
to verify the presence of trend and change points in the data.
If the data fails one of the them at significance level 0.05,
a warning is issued.
}
\examples{

\dontrun{
	## Path the HYDAT database
 db <- DB_HYDAT

 ## Read Amax data
 x <- AmaxData('01AD002', db)

 ## Performing the analysis
 FloodnetAmax(x, period = c(20,50), nsim = 200, verbose = FALSE)
}

}
\references{
Helsel, D. R., & Hirsch, R. M. (2002). Statistical Methods in Water Resources.
   In Techniques of Water-Resources Investigations of the United States
   Geological Survey.
   Retrieved from http://water.usgs.gov/pubs/twri/twri4a3/

Hosking, J. R. M., & Wallis, J. R. (1997). Regional frequency analysis:
  an approach based on L-moments. Cambridge Univ Pr.
}
\seealso{
\link{AmaxData}, \link{floodnetMdl}.
}
