% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetApp.R
\name{FloodnetApp}
\alias{FloodnetApp}
\title{Floodnet App}
\usage{
FloodnetApp()
}
\value{
None
}
\description{
A simple function used to run the Shiny GUI from the floodnetRfa package
}
\examples{

FloodnetToolbox()

}
