% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCA.R, R/SeasonPlot.R, R/floodnetGraphics.R,
%   R/hist_floodnetMdl.R
\name{MapCA}
\alias{MapCA}
\alias{SeasonPlot}
\alias{floodnetGraphics}
\alias{plot.floodnetMdl}
\alias{plot.floodnetMdls}
\alias{hist.floodnetMdl}
\title{Plotting functions for floodnet outputs.}
\usage{
MapCA(..., polygon.args = NULL)

SeasonPlot(..., xlab = "Timing", ylab = "Regularity")

\method{plot}{floodnetMdl}(x, type = "r", ...)

\method{plot}{floodnetMdls}(x, type = "q", ...)

\method{hist}{floodnetMdl}(x, histogram.args = NULL, line.args = NULL, xlab = NULL, ylab = "Density")
}
\arguments{
\item{...}{Other arguments pass to the respective geometry of \code{ggplot2}.
Must be a list of arguments.}

\item{x}{Input data.}

\item{type}{Plot type. \code{'r'} - Return level, \code{'q'} - compare quantile,
\code{'cv'} - compare cv, \code{'l'} - L-moment ratio.}
}
\description{
Generate diagnostic plots based on the output of a floodnet function
\link{floodnetMdl}). Includes the following graphics: Return level plot,
comparison of the flood quantiles, comparison of the coefficient of
variation, L-moment ratio diagram and histogram, seasonal plot and
a basic map of Canada.
}
\details{
All functions use \code{ggplot2} to create a basic plot that can be further
customized. The default arguments of the used geometry can be overide by
passing a list of new arguments. For example,
\code{point.args = list(colour = 'red')}
change the color of a set of points that was created by \code{geom_point}.

The main graphcis are produced by the method \code{plot} (and \code{hist}).
The function \code{SeasonPlot} and \code{MapCA} create basic canvas for
creating a seasonality plot or a map of Canada.
}
\examples{

\dontrun{
x <- rnorm(10)

}


}
\seealso{
\link[ggplot2]{ggplot}, \link[ggplot2]{geom_point},
  \link[ggplot2]{geom_path}, \link[ggplot2]{geom_ribbon}
}
