% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModel.R, R/as_data_frame_floodnetMdl.R,
%   R/print_floodnetMdl.R
\name{CompareModels}
\alias{CompareModels}
\alias{as.data.frame.floodnetMdl}
\alias{as.data.frame.floodnetMdls}
\alias{floodnetMdl}
\alias{print.floodnetMdl}
\alias{print.floodnetMdls}
\title{Floodnet model output}
\usage{
CompareModels(...)

\method{as.data.frame}{floodnetMdl}(x, type = "q", ...)

\method{as.data.frame}{floodnetMdls}(x, type = "q", ...)

\method{print}{floodnetMdl}(x)

\method{print}{floodnetMdls}(x)
}
\description{
All Floodnet functions create a standardized model output that can be used
for printing results or generating graphics.
}
\seealso{
\link{FloodnetAmax},\link{FloodnetPot}, \link{FloodnetPool},
  \link{floodnetGraphics}, \link{hist.floodnetMdl}.
}
