% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPool.R
\name{FloodnetPool}
\alias{FloodnetPool}
\title{Pool frequency analysis using annual maximum floods}
\usage{
FloodnetPool(x, target, period = c(2, 5, 10, 20, 50, 100),
  distr = NULL, tol.H = 2, nsim = 2000, level = 0.95,
  corr = NULL, out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{Hydrometric data of the form: site, year, value.}

\item{target}{Station ID of the target.}

\item{period}{Return periods for which the flood quantiles are estimated.}

\item{distr}{Regional distribution.}

\item{tol.H}{Heterogeneity measure. Stopping criterion representing the
minimal heterogeneity level accepted for a pooling group.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{level}{Confidence level.}

\item{corr}{Coefficient of correlation used for simulation.
The same value is assumed between each pair of sites.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitRegLmom}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be returned.}
}
\description{
Return the results of a regional flood frequency analysis based on annual
maxima and pooling groups.
}
\details{
Estimation is carried out by the L-moment algorithms.
See \link{FitRegLmom} and \link{PoolRemove} for more details.
If not provided the distance between sites is evaluated using the distance
between the regularity and timing of the annual flood peaks.
Confidence intervals and standard deviation are evaluated by a
parametric bootstrap.
An intersite correlation is estimated and used to simulate
from a multivariate normal distribution.
}
\examples{

\dontrun{
## Path the HYDAT database
 db <- DB_HYDAT

 ## Compute distance
 coord <- gaugedSites[,c('lon','lat')]
 rownames(coord) <- gaugedSites$station

 ## Read Amax data
 x <- AmaxData(rownames(coord), db, target = '01AF009',
 							size = 15, distance = dist(coord))

 ## Performing AMAX analysis using L-moments
 FloodnetPool(x, '01AF009', distr = 'gev', period = c(20,50), nsim = 30)

 ## Read Pot data
 info <- gaugedSites[, c('station','auto','area')]
 xd <- DailyPeaksData(info, db, target = '01AF009',
 							size = 15, distance = dist(coord))

 ## Performing POT analysis using L-moments
 FloodnetPool(xd, '01AF009', period = c(20,50), nsim = 30)
}

}
\seealso{
\link{FitRegLmom}, \link{PoolRemove}.
}
