% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPot.R
\name{FloodnetPot}
\alias{FloodnetPot}
\title{Estimating flood quantiles using peaks over threshold}
\usage{
FloodnetPot(
  x,
  u = NULL,
  area = 59874,
  period = c(2, 5, 10, 20, 50, 100),
  nsim = 2000,
  level = 0.95,
  tol.year = 346,
  out.model = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{Hydrometric data. Dataset having three columns: site, date, value.}

\item{u}{Threshold value provided by the user.}

\item{area}{Drainage area. Use to determine the minimal separating time
between flood events. The default value corresponds to 15 days.}

\item{period}{Return period for which the flood quantiles are estimated.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{level}{Confidence level.}

\item{tol.year}{Number of days necessary to consider a year complete. Otherwise the data is removed.}

\item{out.model}{Logical. Should the model be output. This corresponds to the
output of \link{FitPot}.}

\item{verbose}{Logical. Should message and warnings be output.}
}
\description{
Return the flood quantiles, standard deviation, lower and upper bounds
of the confidence interval based on a Peaks Over Threshold (POT) model.
}
\details{
Estimation is carried out by maximum likelihood using the Generalized
Pareto distribution. If not provided, a best candidate threshold is searched.
It corresponds to the lowest
threshold where the goodness-of-fit test of Anderson-Darling has
a p-value greater than 0.25.
In addition, the selected threshold is forced to have (in
average) fewer than 2.5 peaks per year and at least 30 peaks. If
no such threshold is found, the threshold with the maximum p-value is used instead.
The declustering technique for extracting the flood peaks
follows the recommendations of the Water Ressources council of the United States.
In particular, flood peaks must be separated by at least 4 + log(A) days, where
A is the drainage area of the basin in kilometers squared.
Inference on the estimated flood quantiles is performed by parametric bootstraps.

If \code{verbose == TRUE}, a Mann-Kendall test and logistic regression are used
to verify the presence of trends in the mean excess and the number of peaks per years.
If the data fails one of the tests, a warning is issued.
For the logistic regression, an analysis of deviance (F-test) is used to compare the
constant model with polynomial trends.
#' If the data fails one of the tests at significance level 0.05,
a warning is issued.
}
\examples{

 ## Performing the analysis
 x <- DemoData('daily')
 fit <- FloodnetPot(x, u = 1000, area = 14700)
 summary(fit)

}
\references{
Durocher, M., Zadeh, S. M., Burn, D. H., & Ashkar, F. (2018). Comparison of
  automatic procedures for selecting flood peaks over threshold based on
  goodness-of-fit tests. Hydrological Processes, 0(0).
  https://doi.org/10.1002/hyp.13223

Durocher, M., Burn, D. H., & Ashkar, F. (2019). Comparison of estimation
  methods for a nonstationary index-flood model in flood frequency analysis
  using peaks over threshold. Water Resources Research.
  https://doi.org/10.1029/2019WR025305
}
\seealso{
\link{FloodnetAmax}, \link{FloodnetPool}.
}
