% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodnetPot.R
\name{FloodnetPot}
\alias{FloodnetPot}
\title{Estimating flood quantiles using using peaks over threshold}
\usage{
FloodnetPot(site = NULL, db = NULL, x = NULL, period = c(2, 5, 10,
  20, 50, 100), u = NULL, area = NULL, nsim = 2000, alpha = 0.05,
  tol.year = 346, out.model = FALSE, verbose = TRUE)
}
\arguments{
\item{site}{Station ID for HYDAT or a station name for identification.}

\item{db}{File name of the HYDAT database.}

\item{x}{Hydrometric data. Dataset having two columns: Date and value.}

\item{period}{Return period for which the flood quantiles are estimated.}

\item{u}{Threshold value provided by the user.}

\item{area}{Drainage area.}

\item{nsim}{Number of bootstrap samples used for inference.}

\item{alpha}{Probability outside the confidence interval.}

\item{tol.year}{Number of days necessary to consider a year complete.}

\item{out.model}{Logical. Should the model be output. This correspond to the
output of \link{FitPot}. Otherwise only the estimated flood quantiles are
returned.}

\item{verbose}{Logical. Should message and warnings be output.}
}
\description{
Return the flood quantiles, standard deviation, lower and upper bounds
of the confidence interval based on a Peaks Over Threshold (POT) model.
}
\details{
Estimation is carried out by maximum likelihood using the Generalized
Pareto distribution. If not provided, a best candiate threshold is searched.
It corresponds to the lowest
threshold where the goodness-of-fit test of Anderson-Darling has
a p-value greater that 0.25.
In addition, the selected threshold is forced to have (in
average) less than 2.5 peaks per year and at least 30 peaks. If
no threshold is found, the threshold with the maximimum p-value is used instead.
The declustering technique for extracting the flood peaks
follows the recommendations of the Water Ressources council of United States.
In particular, peaks must be separated by at least 4 + log(A) days, where
A is the drainage area of the basin in square kilometers.
Inference on the estimated flood quantiles is performed by parametric bootstrap.

If the HYDAT database is used as input, the the drainage area
is extracted from the database. If no drainage area is provided,
its value is approximated by an empirical relationship with the river mean flow.
This relationship was estimated using the 1114 sites found in the
dataset `gaugedSites`.
It has the form $log(A) <- 4.0934 + 0.9944 * log(M)$, where M is the river mean
flow.

If \code{verbose == TRUE}, a Mann-Kendall test and logistic regression are used
to verify the presence of trends in the mean excess and the number of peaks per years.
If the data fails one of the tests, a warning is issued.
For the logistic regression, an analyse of deviance (F-test) is used to compare the
constant model with polynomial trends.
#' If the data fails one of the tests at significance level 0.05,
a warning is issued.
}
\examples{

\dontrun{
## Assuming that the HYDAT database was already downloaded
db <- "/pathToDB/HYDAT.sqlite"
FloodnetPot(period = c(20,50), site = '01AD003', db = db)
}

}
\references{
Durocher, M., Zadeh, S. M., Burn, D. H., & Ashkar, F. (2018). Comparison of
  automatic procedures for selecting flood peaks over threshold based on
  goodness-of-fit tests. Hydrological Processes, 0(0).
  https://doi.org/10.1002/hyp.13223

Durocher, M., Burn, D. H., & Ashkar, F. (2019). Comparison of estimation
  methods for a nonstationary index-flood model in flood frequency analysis
  using peaks over threshold. Water Resources Research.
  https://doi.org/10.1029/2019WR025305
}
\seealso{
\link{FloodnetAmax}, \link{FloodnetPool}.
}
