% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaugedSites}
\alias{gaugedSites}
\title{Information about 1114 gauged stations in Canada}
\format{An object of class \code{data.frame} with 1114 rows and 29 columns.}
\usage{
gaugedSites
}
\description{
List of stations in HYDAT that has have at least 20 years of observations.
The datasets include descriptors, candidates thresholds for POT, p-values
of trend tests, super regions of hydrologically similar stations.
}
\details{
Apart the station idenfication number (\code{station})
provinces (\code{prov}) and geographical coordinates (\code{lon},\code{lat}),
the dataset includes two other site characteristics : drainage area (\code{area}),
mean annual precipitation (\code{map}).

The super regions are created using clustering techniques: Ward's Method
(\code{hc}) and k-means (\code{km}).
The methods are applied on a subset of 4 standardized variables: \code{lon}, \code{lat},
\code{area} and \code{map}.
Euclidean distance is used to to create groups of 6 and 12 super regions.
For example, the column \code{supreg_km12} have classified the stations in
12 super regions using the k-means method.

Seasonality measures based on the annual maximum events are included in the
dataset.
The regularity (\code{season_radius}) and timing  (\code{season_angle}) of
the annual floods in radian are provided directly and in cartesian coordinates
(\code{season_x, season_y}).

The datasets includes candidates thresholds to help performing POT.
The candidates \code{auto} is the first observations that has a p-value
greater or equal to 0.25 and more than 2.5 peaks per year.
The other suggested threshold are associated with a number of peaks per year (PPY).
For example, \code{ppy175} is associated with 1.75 PPY.

The p-values of trend tests are includes to help identify stationary stations.
For AMAX, the Mann-Kendall (\code{trend_mk}) and the Pettitt's test
(\code{trend_pt}) are provided.
For the mean excess of the POT, the Mann-Kendall test is included and for the
frequency of the peaks the significance of the slope of a logistic regression
model is reported (\code{trend_lg}).

Mostofi Zadeh, S., & Burn, D. H. (2019). A Super Region Approach to
Improve Pooled Flood Frequency Analysis. Canadian Water Resources Journal
/ Revue Canadienne Des Ressources Hydriques, 0(0), 1–14.
https://doi.org/10.1080/07011784.2018.1548946

Durocher, M., Zadeh, S. M., Burn, D. H., & Ashkar, F. (2018). Comparison of
automatic procedures for selecting flood peaks over threshold based on
goodness-of-fit tests. Hydrological Processes, 0(0).
https://doi.org/10.1002/hyp.13223
}
\keyword{datasets}
