% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersiteCorData.R
\name{IntersiteCorData}
\alias{IntersiteCorData}
\title{Intersite correlation matrix for annual maxima}
\usage{
IntersiteCorData(db, x, type = "exp", ...)
}
\arguments{
\item{db}{HYDAT database}

\item{x}{List of stations}

\item{type}{Method used to evaluate the correlation coefficients. One of
\code{'emp'}, \code{'exp'}, \code{'raw'} or \code{'avg'}. See
\link[CSHShydRology]{Intersite} for more details.}

\item{...}{Other parameter passed to \link[CSHShydRology]{Intersite}.}
}
\description{
Return the intersite correlation matrix for a list of stations based on
their annual maximum discharges.
}
\examples{

\dontrun{

## The variable DB_HYDAT contains the path of the HYDAT database.
DB_HYDAT <- '.../Hydat.sqlite3'

## Meta-information about stations
gaugedSites <- read.csv('.../gauged_sites.csv')

sreg <- which(gaugedSites$supreg_km12 == 10)
sites <- as.character(gaugedSites$station)[sreg]

## Intersite correlation for a given super region
x <- IntersiteCorData(DB_HYDAT, sites, smooth = .6, distance.max = 200)

#' ## Intersite correlation for a given super region
x <- IntersiteCorData(DB_HYDAT, sites, type = 'emp' )

}

}
\references{
Durocher, M., Burn, D. H., & Mostofi Zadeh, S. (2018). A nationwide regional
  flood frequency analysis at ungauged sites using ROI/GLS with copulas and
  super regions. Journal of Hydrology, 567, 191–202.
  https://doi.org/10.1016/j.jhydrol.2018.10.011
}
