% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCA.R
\name{MapCA}
\alias{MapCA}
\title{Simple map of Canada}
\usage{
MapCA(..., polygon.args = NULL)
}
\arguments{
\item{...}{Other parameters passed to \link[ggplot2]{ggplot}.}

\item{polygon.args}{Parameters of the displayed polygons.
See \link[ggplot2]{geom_polygon}.}
}
\description{
Return a simple map of Canada using the grammar of graphics \code{ggplot2}.
}
\examples{

library(ggplot2)

MapCA() +
  geom_point(data = gaugedSites, aes(x = lon, y = lat), colour = 'orange')

}
\seealso{
\link{floodnetGraphics}, \link[ggplot2]{geom_polygon}.
}
