% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonPlot.R
\name{SeasonPlot}
\alias{SeasonPlot}
\title{Seasonality plot}
\usage{
SeasonPlot(..., xlab = "Timing", ylab = "Regularity")
}
\arguments{
\item{...}{Other parameters passed to \link{ggplot}.}

\item{xlab, ylab}{Axis labels.}
}
\description{
Create a canvas to represent the timing and the regularity of seasonal measures
in a polar coordinate system.
}
\examples{

library(ggplot2)

SeasonPlot() +
  geom_point(data = gaugedSites,
    aes(x = season_angle, y = season_radius),
    colour = 'orange')


}
